Set up coverage and code quality tools
