--- src/aat/java/uk/gov/hmcts/reform/dg/template_management/performance/NpaSmoke.scala
@@ -1,26 +1,0 @@
-package uk.gov.hmcts.reform.em.orchestrator.performance
-import io.gatling.core.Predef._
-import io.gatling.http.Predef._
-import uk.gov.hmcts.reform.em.orchestrator.testutil.Env
-
-class NpaSmoke extends Simulation with HttpConfiguration {
-
-    object Health {
-
-      val run = exec(http("health")
-        .get(Env.getTestUrl + "/health")
-        .check(
-          status.find.in(200)
-        ))
-        .exec( session => {
-          println( "Response:" )
-          println( session( "RESPONSE_DATA" ).as[String] )
-          session
-      })
-    }
-
-    val scn = scenario("NPA smoke test").exec(Health.run)
-
-    setUp(scn.inject(atOnceUsers(1)).protocols(httpProtocol))
-}
-
