--- src/aat/java/uk/gov/hmcts/reform/dg/template_management/smoke/SmokeTest.java
@@ -1,21 +1,0 @@
-package uk.gov.hmcts.reform.dg.template_management.smoke;
-
-import io.restassured.RestAssured;
-import org.junit.Test;
-import uk.gov.hmcts.reform.dg.template_management.testutil.Env;
-
-public class SmokeTest {
-
-    @Test
-    public void testHealthEndpoint() {
-
-        RestAssured.useRelaxedHTTPSValidation();
-
-        RestAssured.given()
-            .request("GET", Env.getTestUrl() + "/health")
-            .then()
-            .statusCode(200);
-
-
-    }
-}
