+++ src/main/java/uk/gov/hmcts/reform/dg/template/management/appinsights/AppInsightsEvents.java
@@ -1,0 +1,5 @@
+package uk.gov.hmcts.reform.dg.template.management.appinsights;
+
+public enum AppInsightsEvents {
+
+}
