+++ src/main/java/uk/gov/hmcts/reform/dg/template/management/appinsights/EventRepository.java
@@ -1,0 +1,8 @@
+package uk.gov.hmcts.reform.dg.template.management.appinsights;
+
+import java.util.Map;
+
+public interface EventRepository {
+
+    void trackEvent(String name, Map<String, String> properties);
+}
