+++ src/main/java/uk/gov/hmcts/reform/dg/template/management/config/Config.java
@@ -1,0 +1,8 @@
+package uk.gov.hmcts.reform.dg.template.management.config;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class Config {
+
+}
