+++ src/main/java/uk/gov/hmcts/reform/dg/template/management/config/MapperConfiguration.java
@@ -1,0 +1,15 @@
+package uk.gov.hmcts.reform.dg.template.management.config;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class MapperConfiguration {
+
+    @Bean
+    ObjectMapper objectMapper() {
+        return new ObjectMapper();
+    }
+
+}
