+++ src/main/java/uk/gov/hmcts/reform/dg/template/management/config/SwaggerConfiguration.java
@@ -1,0 +1,38 @@
+package uk.gov.hmcts.reform.dg.template.management.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@Configuration
+@EnableSwagger2
+@ComponentScan("uk.gov.hmcts.reform.em.orchestrator.rest")
+public class SwaggerConfiguration {
+
+    private static final String apiVersion = "0.0.1";
+
+    @Bean
+    public Docket api() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(RequestHandlerSelectors.any())
+                .paths(PathSelectors.regex("/api(.*)"))
+                .build()
+                .apiInfo(apiInfo());
+    }
+
+    private ApiInfo apiInfo() {
+        return new ApiInfoBuilder()
+                .title("DG Template Management")
+                .description("API return templates")
+                .version(apiVersion)
+                .build();
+    }
+}
