+++ src/main/java/uk/gov/hmcts/reform/dg/template/management/config/security/AuthCheckerConfiguration.java
@@ -1,0 +1,34 @@
+package uk.gov.hmcts.reform.dg.template.management.config.security;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import javax.servlet.http.HttpServletRequest;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+import java.util.function.Function;
+
+@Configuration
+public class AuthCheckerConfiguration {
+
+    @Value("#{'${authorization.s2s-names-whitelist}'.split(',')}")
+    private List<String> s2sNamesWhiteList;
+
+    @Bean
+    public Function<HttpServletRequest, Collection<String>> authorizedServicesExtractor() {
+        return any -> s2sNamesWhiteList;
+    }
+
+    @Bean
+    public Function<HttpServletRequest, Optional<String>> userIdExtractor() {
+        return request -> Optional.empty();
+    }
+
+    @Bean
+    public Function<HttpServletRequest, Collection<String>> authorizedRolesExtractor() {
+        return request -> Collections.EMPTY_LIST;
+    }
+}
