--- src/main/java/uk/gov/hmcts/reform/dg/template_management/appinsights/AppInsightsEvents.java
@@ -1,5 +1,0 @@
-package uk.gov.hmcts.reform.dg.template_management.appinsights;
-
-public enum AppInsightsEvents {
-
-}
