--- src/main/java/uk/gov/hmcts/reform/dg/template_management/appinsights/EventRepository.java
@@ -1,8 +1,0 @@
-package uk.gov.hmcts.reform.dg.template_management.appinsights;
-
-import java.util.Map;
-
-public interface EventRepository {
-
-    void trackEvent(String name, Map<String, String> properties);
-}
