--- src/main/java/uk/gov/hmcts/reform/dg/template_management/config/Config.java
@@ -1,36 +1,0 @@
-package uk.gov.hmcts.reform.dg.template_management.config;
-
-import okhttp3.OkHttpClient;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.dg.template_management.stitching.StitchingService;
-import uk.gov.hmcts.reform.dg.template_management.stitching.mapper.StitchingDTOMapper;
-
-@Configuration
-public class Config {
-
-    @Autowired
-    private OkHttpClient http;
-
-    @Autowired
-    private AuthTokenGenerator authTokenGenerator;
-
-    @Value("${em-rpa-stitching-api.base-url}")
-    private String stitchingBaseUrl;
-
-    @Value("${em-rpa-stitching-api.resource}")
-    private String stitchingResource;
-
-    @Bean
-    public StitchingService getStitchingService() {
-        return new StitchingService(
-            new StitchingDTOMapper(),
-            http,
-            stitchingBaseUrl + stitchingResource,
-            authTokenGenerator
-        );
-    }
-}
