--- src/main/java/uk/gov/hmcts/reform/dg/template_management/config/Constants.java
@@ -1,14 +1,0 @@
-package uk.gov.hmcts.reform.dg.template_management.config;
-
-/**
- * Application constants.
- */
-public final class Constants {
-
-    public static final String SYSTEM_ACCOUNT = "system";
-    public static final String ANONYMOUS_USER = "anonymoususer";
-    public static final String DEFAULT_LANGUAGE = "en";
-
-    private Constants() {
-    }
-}
