--- src/main/java/uk/gov/hmcts/reform/dg/template_management/config/HttpClientConfiguration.java
@@ -1,15 +1,0 @@
-package uk.gov.hmcts.reform.dg.template_management.config;
-
-import okhttp3.OkHttpClient;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class HttpClientConfiguration {
-
-    @Bean
-    public OkHttpClient okHttpClient(){
-        return new OkHttpClient();
-    }
-
-}
