--- src/main/java/uk/gov/hmcts/reform/dg/template_management/config/MapperConfiguration.java
@@ -1,15 +1,0 @@
-package uk.gov.hmcts.reform.dg.template_management.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class MapperConfiguration {
-
-    @Bean
-    ObjectMapper objectMapper() {
-        return new ObjectMapper();
-    }
-
-}
