--- src/main/java/uk/gov/hmcts/reform/dg/template_management/config/ServiceTokenGeneratorConfiguration.java
@@ -1,26 +1,0 @@
-package uk.gov.hmcts.reform.dg.template_management.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Lazy;
-import uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGeneratorFactory;
-
-@Configuration
-@Lazy
-@EnableFeignClients(basePackageClasses = ServiceAuthorisationApi.class)
-public class ServiceTokenGeneratorConfiguration {
-
-    @Bean
-    public AuthTokenGenerator annotationAppTokenGenerator(
-        @Value("${idam.s2s-auth.totp_secret}") final String secret,
-        @Value("${idam.s2s-auth.microservice}") final String microService,
-        final ServiceAuthorisationApi serviceAuthorisationApi
-    ) {
-        return AuthTokenGeneratorFactory.createDefaultGenerator(secret, microService, serviceAuthorisationApi);
-    }
-
-}
